@@ -29,7 +29,7 @@
     "examples": "yarn workspace example dev",
     "dev": "preconstruct dev",
     "prepare": "husky install",
-    "eslint": "eslint --fix src/**/*.{ts,tsx}",
+    "eslint": "eslint --fix packages/**/src/**/*.{ts,tsx}",
     "test": "jest --coverage",
     "test:watch": "jest --watchAll",
     "typecheck": "tsc --noEmit --emitDeclarationOnly false --strict --jsx react",
@@ -38,44 +38,48 @@
     "vers": "yarn changeset version"
   },
   "devDependencies": {
-    "@babel/core": "7.14.8",
-    "@babel/preset-env": "7.14.8",
-    "@babel/preset-react": "7.14.5",
-    "@babel/preset-typescript": "^7.14.5",
-    "@changesets/changelog-git": "^0.1.7",
-    "@changesets/cli": "^2.16.0",
-    "@preconstruct/cli": "^2.1.0",
-    "@testing-library/react": "^12.0.0",
-    "@types/jest": "^26.0.24",
-    "@types/lodash-es": "^4.17.4",
-    "@types/react": "^17.0.14",
-    "@types/react-dom": "^17.0.9",
+    "@babel/core": "7.17.8",
+    "@babel/preset-env": "7.16.11",
+    "@babel/preset-react": "7.16.7",
+    "@babel/preset-typescript": "^7.16.7",
+    "@changesets/changelog-git": "^0.1.10",
+    "@changesets/cli": "^2.21.1",
+    "@preconstruct/cli": "^2.1.5",
+    "@testing-library/react": "^13.0.0-alpha.5",
+    "@types/jest": "^27.4.1",
+    "@types/react": "^17.0.42",
+    "@types/react-dom": "^17.0.14",
+    "@types/react-native": "0.67.3",
     "@types/react-test-renderer": "^17.0.1",
     "@types/scheduler": "^0.16.2",
-    "@types/three": "^0.133.0",
-    "@typescript-eslint/eslint-plugin": "^4.28.4",
-    "@typescript-eslint/parser": "^4.28.4",
-    "eslint": "^7.31.0",
-    "eslint-config-prettier": "^8.3.0",
+    "@types/three": "^0.138.0",
+    "@typescript-eslint/eslint-plugin": "^5.16.0",
+    "@typescript-eslint/parser": "^5.16.0",
+    "eslint": "^8.11.0",
+    "eslint-config-prettier": "^8.5.0",
     "eslint-import-resolver-alias": "^1.1.2",
-    "eslint-plugin-import": "^2.23.4",
-    "eslint-plugin-jest": "^24.3.6",
-    "eslint-plugin-prettier": "^3.4.0",
-    "eslint-plugin-react": "^7.24.0",
-    "eslint-plugin-react-hooks": "^4.2.0",
-    "husky": "^7.0.1",
-    "jest": "^27.0.6",
-    "jest-cli": "^27.0.6",
-    "lint-staged": "^11.0.1",
-    "prettier": "^2.3.2",
-    "pretty-quick": "^3.1.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "regenerator-runtime": "^0.13.7",
-    "three": "^0.133.0",
-    "three-stdlib": "^2.3.1",
-    "ts-jest": "^27.0.3",
-    "typescript": "^4.3.5"
+    "eslint-plugin-import": "^2.25.4",
+    "eslint-plugin-jest": "^26.1.3",
+    "eslint-plugin-prettier": "^4.0.0",
+    "eslint-plugin-react": "^7.29.4",
+    "eslint-plugin-react-hooks": "^4.3.0",
+    "expo-asset": "^8.4.6",
+    "expo-gl": "^11.1.2",
+    "husky": "^7.0.4",
+    "jest": "^27.5.1",
+    "jest-cli": "^27.5.1",
+    "lint-staged": "^12.3.7",
+    "prettier": "^2.6.0",
+    "pretty-quick": "^3.1.3",
+    "react": "^18.0.0-rc.3",
+    "react-dom": "^18.0.0-rc.3",
+    "react-native": "0.67.4",
+    "react-test-renderer": "^18.0.0-rc.2",
+    "regenerator-runtime": "^0.13.9",
+    "three": "^0.139.0",
+    "three-stdlib": "^2.8.10",
+    "ts-jest": "^27.1.4",
+    "typescript": "^4.6.2"
   },
   "dependencies": {}
 }
