@@ -101,12 +101,7 @@ function Line({ defaultStart, defaultEnd }: any) {
     <>
       <line ref={lineRef as any}>
         <bufferGeometry>
-          <bufferAttribute
-            attachObject={['attributes', 'position']}
-            count={positions.length / 3}
-            array={positions}
-            itemSize={3}
-          />
+          <bufferAttribute attach="attributes-position" count={positions.length / 3} array={positions} itemSize={3} />
         </bufferGeometry>
         <lineBasicMaterial color="white" />
       </line>
