@@ -51,14 +51,9 @@ function Particles({ pointCount }: any) {
 
   return (
     <points ref={points} onPointerOver={hover} onPointerOut={unhover}>
-      <bufferGeometry attach="geometry">
-        <bufferAttribute
-          attachObject={['attributes', 'position']}
-          count={positions.length / 3}
-          array={positions}
-          itemSize={3}
-        />
-        <bufferAttribute attachObject={['attributes', 'color']} count={colors.length / 3} array={colors} itemSize={3} />
+      <bufferGeometry>
+        <bufferAttribute attach="attributes-position" count={positions.length / 3} array={positions} itemSize={3} />
+        <bufferAttribute attach="attributes-color" count={colors.length / 3} array={colors} itemSize={3} />
       </bufferGeometry>
       <dotMaterial vertexColors />
     </points>
