@@ -1,13 +1,20 @@
-import React, { useState, useEffect, useRef } from 'react'
+import * as THREE from 'three'
+import React, { useState, useEffect, useRef, useReducer } from 'react'
 import { Canvas } from '@react-three/fiber'
 
 function Test() {
-  const [vec, setVec] = useState<[number, number, number]>([1, 2, 3])
-  const vecRef = useRef()
-  useEffect(() => void setTimeout(() => setVec([4, 5, 6]), 1000), [])
-  useEffect(() => void console.log('REF ->', vecRef.current), [vec])
-
-  return <vector3 ref={vecRef} args={vec} />
+  const [o1] = useState(
+    () => new THREE.Mesh(new THREE.BoxGeometry(), new THREE.MeshBasicMaterial({ color: 'hotpink' })),
+  )
+  const [o2] = useState(
+    () => new THREE.Mesh(new THREE.BoxGeometry(), new THREE.MeshBasicMaterial({ color: 'aquamarine' })),
+  )
+  const [which, toggle] = useReducer((state) => !state, true)
+  useEffect(() => {
+    const interval = setInterval(toggle, 1000)
+    return () => clearInterval(interval)
+  }, [])
+  return <primitive object={which ? o1 : o2} />
 }
 
 export default function App() {
