@@ -1,5 +1,6 @@
+import * as THREE from 'three'
 import { Root } from './renderer'
-import { RootState } from './store'
+import { RootState, Subscription } from './store'
 
 type GlobalRenderCallback = (timeStamp: number) => void
 
@@ -21,7 +22,8 @@ function run(effects: GlobalRenderCallback[], timestamp: number) {
   for (i = 0; i < effects.length; i++) effects[i](timestamp)
 }
 
-function render(timestamp: number, state: RootState) {
+let subscribers: Subscription[]
+function render(timestamp: number, state: RootState, frame?: THREE.XRFrame) {
   // Run local effects
   let delta = state.clock.getDelta()
   // In frameloop='never' mode, clock times are updated using the provided timestamp
@@ -31,7 +33,8 @@ function render(timestamp: number, state: RootState) {
     state.clock.elapsedTime = timestamp
   }
   // Call subscribers (useFrame)
-  for (i = 0; i < state.internal.subscribers.length; i++) state.internal.subscribers[i].ref.current(state, delta)
+  subscribers = state.internal.subscribers
+  for (i = 0; i < subscribers.length; i++) subscribers[i].ref.current(state, delta, frame)
   // Render content
   if (!state.internal.priority && state.gl.render) state.gl.render(state.scene, state.camera)
   // Decrease frame count
@@ -42,34 +45,46 @@ function render(timestamp: number, state: RootState) {
 export function createLoop<TCanvas>(roots: Map<TCanvas, Root>) {
   let running = false
   let repeat: number
-  function loop(timestamp: number) {
+  let frame: number
+  let state: RootState
+  function loop(timestamp: number): void {
+    frame = requestAnimationFrame(loop)
     running = true
     repeat = 0
 
     // Run effects
-    run(globalEffects, timestamp)
+    if (globalEffects.length) run(globalEffects, timestamp)
+
     // Render all roots
     roots.forEach((root) => {
-      const state = root.store.getState()
+      state = root.store.getState()
       // If the frameloop is invalidated, do not run another frame
-      if (state.internal.active && (state.frameloop === 'always' || state.internal.frames > 0))
+      if (
+        state.internal.active &&
+        (state.frameloop === 'always' || state.internal.frames > 0) &&
+        !state.gl.xr?.isPresenting
+      ) {
         repeat += render(timestamp, state)
+      }
     })
+
     // Run after-effects
-    run(globalAfterEffects, timestamp)
+    if (globalAfterEffects.length) run(globalAfterEffects, timestamp)
 
-    // Keep on looping if anything invalidates the frameloop
-    if (repeat > 0) return requestAnimationFrame(loop)
-    // Tail call effects, they are called when rendering stops
-    else run(globalTailEffects, timestamp)
+    // Stop the loop if nothing invalidates it
+    if (repeat === 0) {
+      // Tail call effects, they are called when rendering stops
+      if (globalTailEffects.length) run(globalTailEffects, timestamp)
 
-    // Flag end of operation
-    running = false
+      // Flag end of operation
+      running = false
+      return cancelAnimationFrame(frame)
+    }
   }
 
   function invalidate(state?: RootState): void {
     if (!state) return roots.forEach((root) => invalidate(root.store.getState()))
-    if (state.vr || !state.internal.active || state.frameloop === 'never') return
+    if (state.gl.xr?.isPresenting || !state.internal.active || state.frameloop === 'never') return
     // Increase frames, do not go higher than 60
     state.internal.frames = Math.min(60, state.internal.frames + 1)
     // If the render-loop isn't active, start it
@@ -79,10 +94,15 @@ export function createLoop<TCanvas>(roots: Map<TCanvas, Root>) {
     }
   }
 
-  function advance(timestamp: number, runGlobalEffects: boolean = true, state?: RootState): void {
+  function advance(
+    timestamp: number,
+    runGlobalEffects: boolean = true,
+    state?: RootState,
+    frame?: THREE.XRFrame,
+  ): void {
     if (runGlobalEffects) run(globalEffects, timestamp)
     if (!state) roots.forEach((root) => render(timestamp, root.store.getState()))
-    else render(timestamp, state)
+    else render(timestamp, state, frame)
     if (runGlobalEffects) run(globalAfterEffects, timestamp)
   }
 
