@@ -1,4 +1,3 @@
-// for type declarations
 export * from './three-types'
 import * as ReactThreeFiber from './three-types'
 export { ReactThreeFiber }
@@ -15,6 +14,7 @@ export type {
   RootState,
 } from './core/store'
 export type { ThreeEvent, Events, EventManager } from './core/events'
-export type { Props } from './web/Canvas'
-
-export * from './web'
+export type { ObjectMap } from './core/utils'
+export * from './web/Canvas'
+export { createPointerEvents as events } from './web/events'
+export * from './core'
