@@ -1,15 +1,12 @@
-jest.mock('scheduler', () => require('scheduler/unstable_mock'))
-
 import * as React from 'react'
-import { Group, Camera, Scene, Raycaster, Mesh, BoxBufferGeometry, MeshBasicMaterial } from 'three'
-// @ts-ignore
+import * as THREE from 'three'
 import * as Stdlib from 'three-stdlib'
 import { createCanvas } from '@react-three/test-renderer/src/createTestCanvas'
 import { createWebGLContext } from '@react-three/test-renderer/src/createWebGLContext'
 
 import { asyncUtils } from '../../../shared/asyncUtils'
 
-import { render, advance, useLoader, act, useThree, useGraph, useFrame, ObjectMap } from '../../src/web/index'
+import { createRoot, advance, useLoader, act, useThree, useGraph, useFrame, ObjectMap } from '../../src'
 
 const resolvers = []
 
@@ -17,7 +14,7 @@ const { waitFor } = asyncUtils(act, (resolver: () => void) => {
   resolvers.push(resolver)
 })
 
-describe('web hooks', () => {
+describe('hooks', () => {
   let canvas: HTMLCanvasElement = null!
 
   beforeEach(() => {
@@ -35,9 +32,9 @@ describe('web hooks', () => {
 
   it('can handle useThree hook', async () => {
     let result = {} as {
-      camera: Camera
-      scene: Scene
-      raycaster: Raycaster
+      camera: THREE.Camera
+      scene: THREE.Scene
+      raycaster: THREE.Raycaster
       size: { width: number; height: number }
     }
 
@@ -60,12 +57,12 @@ describe('web hooks', () => {
     }
 
     await act(async () => {
-      render(<Component />, canvas)
+      createRoot(canvas).render(<Component />)
     })
 
-    expect(result.camera instanceof Camera).toBeTruthy()
-    expect(result.scene instanceof Scene).toBeTruthy()
-    expect(result.raycaster instanceof Raycaster).toBeTruthy()
+    expect(result.camera instanceof THREE.Camera).toBeTruthy()
+    expect(result.scene instanceof THREE.Scene).toBeTruthy()
+    expect(result.raycaster instanceof THREE.Raycaster).toBeTruthy()
     expect(result.size).toEqual({ height: 0, width: 0 })
   })
 
@@ -87,15 +84,21 @@ describe('web hooks', () => {
       )
     }
 
-    let scene: Scene = null!
-    await act(async () => (scene = render(<Component />, canvas, { frameloop: 'never' }).getState().scene))
+    let scene: THREE.Scene = null!
+    await act(
+      async () =>
+        (scene = createRoot(canvas)
+          .configure({ frameloop: 'never' })
+          .render(<Component />)
+          .getState().scene),
+    )
     advance(Date.now())
     expect(scene.children[0].position.x).toEqual(1)
     expect(frameCalls.length).toBeGreaterThan(0)
   })
 
   it('can handle useLoader hook', async () => {
-    const MockMesh = new Mesh()
+    const MockMesh = new THREE.Mesh()
     // @ts-ignore
     jest.spyOn(Stdlib, 'GLTFLoader').mockImplementation(() => ({
       load: jest.fn().mockImplementation((url, onLoad) => {
@@ -109,14 +112,15 @@ describe('web hooks', () => {
       return <primitive object={model} />
     }
 
-    let scene: Scene = null!
+    let scene: THREE.Scene = null!
     await act(async () => {
-      scene = render(
-        <React.Suspense fallback={null}>
-          <Component />
-        </React.Suspense>,
-        canvas,
-      ).getState().scene
+      scene = createRoot(canvas)
+        .render(
+          <React.Suspense fallback={null}>
+            <Component />
+          </React.Suspense>,
+        )
+        .getState().scene
     })
 
     await waitFor(() => expect(scene.children[0]).toBeDefined())
@@ -125,16 +129,16 @@ describe('web hooks', () => {
   })
 
   it('can handle useLoader hook with an array of strings', async () => {
-    const MockMesh = new Mesh()
+    const MockMesh = new THREE.Mesh()
 
-    const MockGroup = new Group()
-    const mat1 = new MeshBasicMaterial()
+    const MockGroup = new THREE.Group()
+    const mat1 = new THREE.MeshBasicMaterial()
     mat1.name = 'Mat 1'
-    const mesh1 = new Mesh(new BoxBufferGeometry(2, 2), mat1)
+    const mesh1 = new THREE.Mesh(new THREE.BoxBufferGeometry(2, 2), mat1)
     mesh1.name = 'Mesh 1'
-    const mat2 = new MeshBasicMaterial()
+    const mat2 = new THREE.MeshBasicMaterial()
     mat2.name = 'Mat 2'
-    const mesh2 = new Mesh(new BoxBufferGeometry(2, 2), mat2)
+    const mesh2 = new THREE.Mesh(new THREE.BoxBufferGeometry(2, 2), mat2)
     mesh2.name = 'Mesh 2'
     MockGroup.add(mesh1, mesh2)
 
@@ -166,14 +170,15 @@ describe('web hooks', () => {
       )
     }
 
-    let scene: Scene = null!
+    let scene: THREE.Scene = null!
     await act(async () => {
-      scene = render(
-        <React.Suspense fallback={null}>
-          <Component />
-        </React.Suspense>,
-        canvas,
-      ).getState().scene
+      scene = createRoot(canvas)
+        .render(
+          <React.Suspense fallback={null}>
+            <Component />
+          </React.Suspense>,
+        )
+        .getState().scene
     })
 
     await waitFor(() => expect(scene.children[0]).toBeDefined())
@@ -182,23 +187,23 @@ describe('web hooks', () => {
   })
 
   it('can handle useGraph hook', async () => {
-    const group = new Group()
-    const mat1 = new MeshBasicMaterial()
+    const group = new THREE.Group()
+    const mat1 = new THREE.MeshBasicMaterial()
     mat1.name = 'Mat 1'
-    const mesh1 = new Mesh(new BoxBufferGeometry(2, 2), mat1)
+    const mesh1 = new THREE.Mesh(new THREE.BoxBufferGeometry(2, 2), mat1)
     mesh1.name = 'Mesh 1'
-    const mat2 = new MeshBasicMaterial()
+    const mat2 = new THREE.MeshBasicMaterial()
     mat2.name = 'Mat 2'
-    const mesh2 = new Mesh(new BoxBufferGeometry(2, 2), mat2)
+    const mesh2 = new THREE.Mesh(new THREE.BoxBufferGeometry(2, 2), mat2)
     mesh2.name = 'Mesh 2'
-    const subGroup = new Group()
-    const mat3 = new MeshBasicMaterial()
+    const subGroup = new THREE.Group()
+    const mat3 = new THREE.MeshBasicMaterial()
     mat3.name = 'Mat 3'
-    const mesh3 = new Mesh(new BoxBufferGeometry(2, 2), mat3)
+    const mesh3 = new THREE.Mesh(new THREE.BoxBufferGeometry(2, 2), mat3)
     mesh3.name = 'Mesh 3'
-    const mat4 = new MeshBasicMaterial()
+    const mat4 = new THREE.MeshBasicMaterial()
     mat4.name = 'Mat 4'
-    const mesh4 = new Mesh(new BoxBufferGeometry(2, 2), mat4)
+    const mesh4 = new THREE.Mesh(new THREE.BoxBufferGeometry(2, 2), mat4)
     mesh4.name = 'Mesh 4'
 
     subGroup.add(mesh3, mesh4)
@@ -213,7 +218,7 @@ describe('web hooks', () => {
     }
 
     await act(async () => {
-      render(<Component />, canvas)
+      createRoot(canvas).render(<Component />)
     })
 
     expect(result).toEqual({
