@@ -1,48 +1,30 @@
-jest.mock('scheduler', () => require('scheduler/unstable_mock'))
-
 import * as React from 'react'
-import {
-  Color,
-  Group,
-  Camera,
-  Scene,
-  Mesh,
-  BoxBufferGeometry,
-  MeshBasicMaterial,
-  MeshStandardMaterial,
-  PCFSoftShadowMap,
-  ACESFilmicToneMapping,
-  sRGBEncoding,
-  Object3D,
-  WebGLRenderer,
-  LinearEncoding,
-  NoToneMapping,
-} from 'three'
+import * as THREE from 'three'
 import { createCanvas } from '@react-three/test-renderer/src/createTestCanvas'
 import { createWebGLContext } from '@react-three/test-renderer/src/createWebGLContext'
 
-import { render, act, unmountComponentAtNode, extend } from '../../src/web/index'
-import { UseBoundStore } from 'zustand'
+import { createRoot, act, unmountComponentAtNode, useFrame, extend, ReactThreeFiber } from '../../src/index'
+import { UseStore } from 'zustand'
 import { RootState } from '../../src/core/store'
-import { ReactThreeFiber } from '../../src'
+import { Instance } from '../../src/core/renderer'
 
-type ComponentMesh = Mesh<BoxBufferGeometry, MeshBasicMaterial>
+type ComponentMesh = THREE.Mesh<THREE.BoxBufferGeometry, THREE.MeshBasicMaterial>
 
 /* This class is used for one of the tests */
-class HasObject3dMember extends Object3D {
-  public attachment?: Object3D = undefined
+class HasObject3dMember extends THREE.Object3D {
+  public attachment?: THREE.Object3D = undefined
 }
 
 /* This class is used for one of the tests */
-class HasObject3dMethods extends Object3D {
-  attachedObj3d?: Object3D
-  detachedObj3d?: Object3D
+class HasObject3dMethods extends THREE.Object3D {
+  attachedObj3d?: THREE.Object3D
+  detachedObj3d?: THREE.Object3D
 
-  customAttach(obj3d: Object3D) {
+  customAttach(obj3d: THREE.Object3D) {
     this.attachedObj3d = obj3d
   }
 
-  detach(obj3d: Object3D) {
+  detach(obj3d: THREE.Object3D) {
     this.detachedObj3d = obj3d
   }
 }
@@ -65,7 +47,7 @@ beforeAll(() => {
   })
 })
 
-describe('web core', () => {
+describe('renderer', () => {
   let canvas: HTMLCanvasElement = null!
 
   beforeEach(() => {
@@ -90,24 +72,28 @@ describe('web core', () => {
         </mesh>
       )
     }
-    let scene: Scene = null!
+    let scene: THREE.Scene = null!
     await act(async () => {
-      scene = render(<Mesh />, canvas).getState().scene
+      scene = createRoot(canvas)
+        .render(<Mesh />)
+        .getState().scene
     })
 
     expect(scene.children[0].type).toEqual('Mesh')
     expect((scene.children[0] as ComponentMesh).geometry.type).toEqual('BoxGeometry')
     expect((scene.children[0] as ComponentMesh).material.type).toEqual('MeshBasicMaterial')
-    expect((scene.children[0] as THREE.Mesh<THREE.BoxGeometry, MeshStandardMaterial>).material.type).toEqual(
+    expect((scene.children[0] as THREE.Mesh<THREE.BoxGeometry, THREE.MeshStandardMaterial>).material.type).toEqual(
       'MeshBasicMaterial',
     )
   })
 
   it('renders an empty scene', async () => {
     const Empty = () => null
-    let scene: Scene = null!
+    let scene: THREE.Scene = null!
     await act(async () => {
-      scene = render(<Empty />, canvas).getState().scene
+      scene = createRoot(canvas)
+        .render(<Empty />)
+        .getState().scene
     })
 
     expect(scene.type).toEqual('Scene')
@@ -135,19 +121,21 @@ describe('web core', () => {
       )
     }
 
-    let scene: Scene = null!
+    let scene: THREE.Scene = null!
     await act(async () => {
-      scene = render(<Parent />, canvas).getState().scene
+      scene = createRoot(canvas)
+        .render(<Parent />)
+        .getState().scene
     })
 
     expect(scene.children[0].type).toEqual('Group')
-    // @ts-expect-error we do append background to group, but it's not wrong because it won't do anything.
+    // @ts-ignore we do append background to group, but it's not wrong because it won't do anything.
     expect((scene.children[0] as Group).background.getStyle()).toEqual('rgb(0,0,0)')
     expect(scene.children[0].children[0].type).toEqual('Mesh')
     expect((scene.children[0].children[0] as ComponentMesh).geometry.type).toEqual('BoxGeometry')
     expect((scene.children[0].children[0] as ComponentMesh).material.type).toEqual('MeshBasicMaterial')
     expect(
-      (scene.children[0].children[0] as THREE.Mesh<THREE.BoxGeometry, MeshStandardMaterial>).material.type,
+      (scene.children[0].children[0] as THREE.Mesh<THREE.BoxGeometry, THREE.MeshStandardMaterial>).material.type,
     ).toEqual('MeshBasicMaterial')
   })
 
@@ -182,61 +170,97 @@ describe('web core', () => {
       return null
     }
 
-    let scene: Scene = null!
+    let scene: THREE.Scene = null!
     await act(async () => {
-      scene = render(<Component />, canvas).getState().scene
+      scene = createRoot(canvas)
+        .render(<Component />)
+        .getState().scene
     })
 
     expect(scene.children[0].position.x).toEqual(7)
-    expect(renders).toBe(12)
+    expect(renders).toBe(6)
   })
 
   it('updates types & names', async () => {
-    let scene: Scene = null!
+    let scene: THREE.Scene = null!
     await act(async () => {
-      scene = render(
-        <mesh>
-          <meshBasicMaterial name="basicMat">
-            <color attach="color" args={[0, 0, 0]} />
-          </meshBasicMaterial>
-        </mesh>,
-        canvas,
-      ).getState().scene
+      scene = createRoot(canvas)
+        .render(
+          <mesh>
+            <meshBasicMaterial name="basicMat">
+              <color attach="color" args={[0, 0, 0]} />
+            </meshBasicMaterial>
+          </mesh>,
+        )
+        .getState().scene
     })
 
-    expect((scene.children[0] as THREE.Mesh<THREE.BoxGeometry, MeshBasicMaterial>).material.type).toEqual(
+    expect((scene.children[0] as THREE.Mesh<THREE.BoxGeometry, THREE.MeshBasicMaterial>).material.type).toEqual(
       'MeshBasicMaterial',
     )
-    expect((scene.children[0] as THREE.Mesh<THREE.BoxGeometry, MeshBasicMaterial>).material.name).toEqual('basicMat')
+    expect((scene.children[0] as THREE.Mesh<THREE.BoxGeometry, THREE.MeshBasicMaterial>).material.name).toEqual(
+      'basicMat',
+    )
 
     await act(async () => {
-      scene = render(
-        <mesh>
-          <meshStandardMaterial name="standardMat">
-            <color attach="color" args={[255, 255, 255]} />
-          </meshStandardMaterial>
-        </mesh>,
-        canvas,
-      ).getState().scene
+      scene = createRoot(canvas)
+        .render(
+          <mesh>
+            <meshStandardMaterial name="standardMat">
+              <color attach="color" args={[255, 255, 255]} />
+            </meshStandardMaterial>
+          </mesh>,
+        )
+        .getState().scene
     })
 
-    expect((scene.children[0] as THREE.Mesh<THREE.BoxGeometry, MeshStandardMaterial>).material.type).toEqual(
+    expect((scene.children[0] as THREE.Mesh<THREE.BoxGeometry, THREE.MeshStandardMaterial>).material.type).toEqual(
       'MeshStandardMaterial',
     )
-    expect((scene.children[0] as THREE.Mesh<THREE.BoxGeometry, MeshStandardMaterial>).material.name).toEqual(
+    expect((scene.children[0] as THREE.Mesh<THREE.BoxGeometry, THREE.MeshStandardMaterial>).material.name).toEqual(
       'standardMat',
     )
   })
 
-  it('attaches Object3D children that use attach', async () => {
-    let scene: Scene = null!
+  it('should forward ref three object', async () => {
+    // Note: Passing directly should be less strict, and assigning current should be more strict
+    let immutableRef!: React.RefObject<THREE.Mesh>
+    let mutableRef!: React.MutableRefObject<THREE.Object3D | undefined>
+    let mutableRefSpecific!: React.MutableRefObject<THREE.Mesh | null>
+
+    const RefTest = () => {
+      immutableRef = React.createRef()
+      mutableRef = React.useRef()
+      mutableRefSpecific = React.useRef(null)
+
+      return (
+        <>
+          <mesh ref={immutableRef} />
+          <mesh ref={mutableRef} />
+          <mesh ref={(r) => (mutableRefSpecific.current = r)} />
+        </>
+      )
+    }
+
     await act(async () => {
-      scene = render(
-        <hasObject3dMember>
-          <mesh attach="attachment" />
-        </hasObject3dMember>,
-        canvas,
-      ).getState().scene
+      createRoot(canvas).render(<RefTest />)
+    })
+
+    expect(immutableRef.current).toBeTruthy()
+    expect(mutableRef.current).toBeTruthy()
+    expect(mutableRefSpecific.current).toBeTruthy()
+  })
+
+  it('attaches Object3D children that use attach', async () => {
+    let scene: THREE.Scene = null!
+    await act(async () => {
+      scene = createRoot(canvas)
+        .render(
+          <hasObject3dMember>
+            <mesh attach="attachment" />
+          </hasObject3dMember>,
+        )
+        .getState().scene
     })
 
     const attachedMesh = (scene.children[0] as HasObject3dMember).attachment
@@ -247,14 +271,15 @@ describe('web core', () => {
   })
 
   it('can attach a Scene', async () => {
-    let scene: Scene = null!
+    let scene: THREE.Scene = null!
     await act(async () => {
-      scene = render(
-        <hasObject3dMember>
-          <scene attach="attachment" />
-        </hasObject3dMember>,
-        canvas,
-      ).getState().scene
+      scene = createRoot(canvas)
+        .render(
+          <hasObject3dMember>
+            <scene attach="attachment" />
+          </hasObject3dMember>,
+        )
+        .getState().scene
     })
 
     const attachedScene = (scene.children[0] as HasObject3dMember).attachment
@@ -265,15 +290,16 @@ describe('web core', () => {
   })
 
   describe('attaches Object3D children that use attachFns', () => {
-    it('attachFns as strings', async () => {
-      let scene: Scene = null!
+    it('attachFns with cleanup', async () => {
+      let scene: THREE.Scene = null!
       await act(async () => {
-        scene = render(
-          <hasObject3dMethods>
-            <mesh attachFns={['customAttach', 'detach']} />
-          </hasObject3dMethods>,
-          canvas,
-        ).getState().scene
+        scene = createRoot(canvas)
+          .render(
+            <hasObject3dMethods>
+              <mesh attach={(parent, self) => (parent.customAttach(self), () => parent.detach(self))} />
+            </hasObject3dMethods>,
+          )
+          .getState().scene
       })
 
       const attachedMesh = (scene.children[0] as HasObject3dMethods).attachedObj3d
@@ -286,7 +312,7 @@ describe('web core', () => {
       expect((scene.children[0] as HasObject3dMethods).detachedObj3d).toBeUndefined()
 
       await act(async () => {
-        render(<hasObject3dMethods />, canvas)
+        createRoot(canvas).render(<hasObject3dMethods />)
       })
 
       const detachedMesh = (scene.children[0] as HasObject3dMethods).detachedObj3d
@@ -294,35 +320,27 @@ describe('web core', () => {
     })
 
     it('attachFns as functions', async () => {
-      let scene: Scene = null!
-      let attachedMesh: Mesh = null!
-      let detachedMesh: Mesh = null!
+      let scene: THREE.Scene = null!
+      let attachedMesh: Instance = null!
+      let detachedMesh: Instance = null!
 
       await act(async () => {
-        scene = render(
-          <hasObject3dMethods>
-            <mesh
-              attachFns={[
-                (mesh: Mesh) => {
-                  attachedMesh = mesh
-                },
-                (mesh: Mesh) => {
-                  detachedMesh = mesh
-                },
-              ]}
-            />
-          </hasObject3dMethods>,
-          canvas,
-        ).getState().scene
+        scene = createRoot(canvas)
+          .render(
+            <hasObject3dMethods>
+              <mesh attach={(parent) => ((attachedMesh = parent), () => (detachedMesh = parent))} />
+            </hasObject3dMethods>,
+          )
+          .getState().scene
       })
 
       expect(attachedMesh).toBeDefined()
-      expect(attachedMesh?.type).toBe('Mesh')
+      expect(attachedMesh?.type).toBe('Object3D')
       // attaching is *instead of* being a regular child
       expect(scene.children[0].children.length).toBe(0)
 
       await act(async () => {
-        render(<hasObject3dMethods />, canvas)
+        createRoot(canvas).render(<hasObject3dMethods />)
       })
 
       expect(detachedMesh).toBe(attachedMesh)
@@ -345,14 +363,14 @@ describe('web core', () => {
     }
 
     await act(async () => {
-      render(<Log key="foo" name="Foo" />, canvas)
+      createRoot(canvas).render(<Log key="foo" name="Foo" />)
     })
 
     await act(async () => {
       unmountComponentAtNode(canvas)
     })
 
-    expect(log).toEqual(['render Foo', 'render Foo', 'mount Foo', 'unmount Foo'])
+    expect(log).toEqual(['render Foo', 'mount Foo', 'unmount Foo'])
   })
 
   it('will mount/unmount event handlers correctly', async () => {
@@ -365,34 +383,44 @@ describe('web core', () => {
     // Test initial mount without events
     mounted = true
     await act(async () => {
-      state = render(<EventfulComponent />, canvas).getState()
+      state = createRoot(canvas)
+        .render(<EventfulComponent />)
+        .getState()
     })
     expect(state.internal.interaction.length).toBe(0)
 
     // Test initial mount with events
     attachEvents = true
     await act(async () => {
-      state = render(<EventfulComponent />, canvas).getState()
+      state = createRoot(canvas)
+        .render(<EventfulComponent />)
+        .getState()
     })
     expect(state.internal.interaction.length).not.toBe(0)
 
     // Test events update
     attachEvents = false
     await act(async () => {
-      state = render(<EventfulComponent />, canvas).getState()
+      state = createRoot(canvas)
+        .render(<EventfulComponent />)
+        .getState()
     })
     expect(state.internal.interaction.length).toBe(0)
 
     attachEvents = true
     await act(async () => {
-      state = render(<EventfulComponent />, canvas).getState()
+      state = createRoot(canvas)
+        .render(<EventfulComponent />)
+        .getState()
     })
     expect(state.internal.interaction.length).not.toBe(0)
 
     // Test unmount with events
     mounted = false
     await act(async () => {
-      state = render(<EventfulComponent />, canvas).getState()
+      state = createRoot(canvas)
+        .render(<EventfulComponent />)
+        .getState()
     })
     expect(state.internal.interaction.length).toBe(0)
   })
@@ -402,14 +430,16 @@ describe('web core', () => {
     const instances: { uuid: string; parentUUID?: string; childUUID?: string }[] = []
 
     const Test = ({ n }: { n: number }) => (
-      // @ts-expect-error args isn't a valid prop but changing it will swap
+      // @ts-ignore args isn't a valid prop but changing it will swap
       <group args={[n]} onPointerOver={() => null}>
         <group />
       </group>
     )
 
     await act(async () => {
-      state = render(<Test n={1} />, canvas).getState()
+      state = createRoot(canvas)
+        .render(<Test n={1} />)
+        .getState()
     })
 
     instances.push({
@@ -419,7 +449,9 @@ describe('web core', () => {
     })
 
     await act(async () => {
-      state = render(<Test n={2} />, canvas).getState()
+      state = createRoot(canvas)
+        .render(<Test n={2} />)
+        .getState()
     })
 
     instances.push({
@@ -442,10 +474,13 @@ describe('web core', () => {
   })
 
   it('will make an Orthographic Camera & set the position', async () => {
-    let camera: Camera = null!
+    let camera: THREE.Camera = null!
 
     await act(async () => {
-      camera = render(<group />, canvas, { orthographic: true, camera: { position: [0, 0, 5] } }).getState().camera
+      camera = createRoot(canvas)
+        .configure({ orthographic: true, camera: { position: [0, 0, 5] } })
+        .render(<group />)
+        .getState().camera
     })
 
     expect(camera.type).toEqual('OrthographicCamera')
@@ -453,12 +488,11 @@ describe('web core', () => {
   })
 
   it('should handle an performance changing functions', async () => {
-    let state: UseBoundStore<RootState> = null!
+    let state: UseStore<RootState> = null!
     await act(async () => {
-      state = render(<group />, canvas, {
-        dpr: [1, 2],
-        performance: { min: 0.2 },
-      })
+      state = createRoot(canvas)
+        .configure({ dpr: [1, 2], performance: { min: 0.2 } })
+        .render(<group />)
     })
 
     expect(state.getState().viewport.initialDpr).toEqual(2)
@@ -490,64 +524,108 @@ describe('web core', () => {
   })
 
   it('should set PCFSoftShadowMap as the default shadow map', async () => {
-    let state: UseBoundStore<RootState> = null!
+    let state: UseStore<RootState> = null!
     await act(async () => {
-      state = render(<group />, canvas, {
-        shadows: true,
-      })
+      state = createRoot(canvas)
+        .configure({ shadows: true })
+        .render(<group />)
     })
 
-    expect(state.getState().gl.shadowMap.type).toBe(PCFSoftShadowMap)
+    expect(state.getState().gl.shadowMap.type).toBe(THREE.PCFSoftShadowMap)
   })
 
   it('should set tonemapping to ACESFilmicToneMapping and outputEncoding to sRGBEncoding if linear is false', async () => {
-    let state: UseBoundStore<RootState> = null!
+    let state: UseStore<RootState> = null!
     await act(async () => {
-      state = render(<group />, canvas, {
-        linear: false,
-      })
+      state = createRoot(canvas)
+        .configure({ linear: false })
+        .render(<group />)
     })
 
-    expect(state.getState().gl.toneMapping).toBe(ACESFilmicToneMapping)
-    expect(state.getState().gl.outputEncoding).toBe(sRGBEncoding)
+    expect(state.getState().gl.toneMapping).toBe(THREE.ACESFilmicToneMapping)
+    expect(state.getState().gl.outputEncoding).toBe(THREE.sRGBEncoding)
+  })
+
+  it('should toggle render mode in xr', async () => {
+    let state: RootState = null!
+
+    await act(async () => {
+      state = createRoot(canvas)
+        .render(<group />)
+        .getState()
+      state.gl.xr.isPresenting = true
+      state.gl.xr.dispatchEvent({ type: 'sessionstart' })
+    })
+
+    expect(state.gl.xr.enabled).toEqual(true)
+
+    await act(async () => {
+      state.gl.xr.isPresenting = false
+      state.gl.xr.dispatchEvent({ type: 'sessionend' })
+    })
+
+    expect(state.gl.xr.enabled).toEqual(false)
+  })
+
+  it('should respect frameloop="never" in xr', async () => {
+    let respected = true
+
+    await act(async () => {
+      const TestGroup = () => {
+        useFrame(() => (respected = false))
+        return <group />
+      }
+      const state = createRoot(canvas)
+        .configure({ frameloop: 'never' })
+        .render(<TestGroup />)
+        .getState()
+      state.gl.xr.isPresenting = true
+      state.gl.xr.dispatchEvent({ type: 'sessionstart' })
+    })
+
+    expect(respected).toEqual(true)
   })
 
   it('will render components that are extended', async () => {
-    class MyColor extends Color {
+    class MyColor extends THREE.Color {
       constructor(col: number) {
         super(col)
       }
     }
 
-    expect(() => {
-      act(async () => {
+    const testExtend = async () => {
+      await act(async () => {
         extend({ MyColor })
 
-        // @ts-expect-error we're testing the extend feature, i'm not adding it to the namespace
-        render(<myColor args={[0x0000ff]} />, canvas)
+        // @ts-ignore we're testing the extend feature, i'm not adding it to the namespace
+        createRoot(canvas).render(<myColor args={[0x0000ff]} />)
       })
-    }).not.toThrow()
+    }
+
+    expect(() => testExtend()).not.toThrow()
   })
 
   it('should set renderer props via gl prop', async () => {
     let gl: THREE.WebGLRenderer = null!
     await act(async () => {
-      gl = render(<group />, canvas, {
-        gl: { physicallyCorrectLights: true },
-      }).getState().gl
+      gl = createRoot(canvas)
+        .configure({ gl: { physicallyCorrectLights: true } })
+        .render(<group />)
+        .getState().gl
     })
 
     expect(gl.physicallyCorrectLights).toBe(true)
   })
 
   it('should set a renderer via gl callback', async () => {
-    class Renderer extends WebGLRenderer {}
+    class Renderer extends THREE.WebGLRenderer {}
 
     let gl: Renderer = null!
     await act(async () => {
-      gl = render(<group />, canvas, {
-        gl: (canvas) => new Renderer({ canvas }),
-      }).getState().gl
+      gl = createRoot(canvas)
+        .configure({ gl: (canvas) => new Renderer({ canvas }) })
+        .render(<group />)
+        .getState().gl
     })
 
     expect(gl instanceof Renderer).toBe(true)
@@ -556,22 +634,22 @@ describe('web core', () => {
   it('should respect color management preferences via gl', async () => {
     let gl: THREE.WebGLRenderer = null!
     await act(async () => {
-      gl = render(<group />, canvas, {
-        gl: { outputEncoding: LinearEncoding, toneMapping: NoToneMapping },
-      }).getState().gl
+      gl = createRoot(canvas)
+        .configure({ gl: { outputEncoding: THREE.LinearEncoding, toneMapping: THREE.NoToneMapping } })
+        .render(<group />)
+        .getState().gl
     })
 
-    expect(gl.outputEncoding).toBe(LinearEncoding)
-    expect(gl.toneMapping).toBe(NoToneMapping)
+    expect(gl.outputEncoding).toBe(THREE.LinearEncoding)
+    expect(gl.toneMapping).toBe(THREE.NoToneMapping)
 
     await act(async () => {
-      gl = render(<group />, canvas, {
-        flat: true,
-        linear: true,
-      }).getState().gl
+      gl = createRoot(canvas)
+        .configure({ flat: true, linear: true })
+        .render(<group />)
+        .getState().gl
     })
-
-    expect(gl.outputEncoding).toBe(LinearEncoding)
-    expect(gl.toneMapping).toBe(NoToneMapping)
+    expect(gl.outputEncoding).toBe(THREE.LinearEncoding)
+    expect(gl.toneMapping).toBe(THREE.NoToneMapping)
   })
 })
