@@ -1,12 +1,7 @@
 import * as React from 'react'
+import * as THREE from 'three'
 
-import {
-  _roots as mockRoots,
-  render,
-  reconciler,
-  unmountComponentAtNode as unmount,
-  act as _act,
-} from '@react-three/fiber'
+import { extend, _roots as mockRoots, createRoot, reconciler, act as _act } from '@react-three/fiber'
 
 import { toTree } from './helpers/tree'
 import { toGraph } from './helpers/graph'
@@ -19,6 +14,11 @@ import { createEventFirer } from './fireEvent'
 import type { MockScene } from './types/internal'
 import type { CreateOptions, Renderer, Act } from './types/public'
 import { wrapFiber } from './createTestInstance'
+
+// Extend catalogue for render API in tests.
+extend(THREE)
+
+const act = _act as unknown as Act
 
 const create = async (element: React.ReactNode, options?: Partial<CreateOptions>): Promise<Renderer> => {
   const canvas = createCanvas({
@@ -36,11 +36,12 @@ const create = async (element: React.ReactNode, options?: Partial<CreateOptions>
 
   const _fiber = canvas
 
+  const _root = createRoot(_fiber).configure({ frameloop: 'never', ...options, events: undefined })
+
   let scene: MockScene = null!
 
-  await reconciler.act(async () => {
-    scene = (render(element, _fiber, { frameloop: 'never', ...options, events: undefined }).getState()
-      .scene as unknown) as MockScene
+  await act(async () => {
+    scene = _root.render(element).getState().scene as unknown as MockScene
   })
 
   const _store = mockRoots.get(_fiber)!.store
@@ -48,8 +49,8 @@ const create = async (element: React.ReactNode, options?: Partial<CreateOptions>
   return {
     scene: wrapFiber(scene),
     unmount: async () => {
-      await reconciler.act(async () => {
-        unmount(_fiber)
+      await act(async () => {
+        _root.unmount()
       })
     },
     getInstance: () => {
@@ -78,7 +79,7 @@ const create = async (element: React.ReactNode, options?: Partial<CreateOptions>
     update: async (newElement: React.ReactNode) => {
       const fiber = mockRoots.get(_fiber)?.fiber
       if (fiber) {
-        await reconciler.act(async () => {
+        await act(async () => {
           reconciler.updateContainer(newElement, fiber, null, () => null)
         })
       }
@@ -90,7 +91,7 @@ const create = async (element: React.ReactNode, options?: Partial<CreateOptions>
     toGraph: () => {
       return toGraph(scene)
     },
-    fireEvent: createEventFirer(reconciler.act, _store),
+    fireEvent: createEventFirer(act, _store),
     advanceFrames: async (frames: number, delta: number | number[] = 1) => {
       const state = _store.getState()
       const storeSubscribers = state.internal.subscribers
@@ -113,8 +114,6 @@ const create = async (element: React.ReactNode, options?: Partial<CreateOptions>
     },
   }
 }
-
-const act = (_act as unknown) as Act
 
 export * as ReactThreeTest from './types'
 export default { create, act }
